Fix: Computer
1. Computer Entity 내에는 실행 메서드를 작성하지 않고, 이를 Service 로 모아서 작성
2. Service 에는 숫자랜덤생성, 재생성, 비교, 스트라이크/볼 카운트, 프린트 를 메서드로 나누어서 작성
