+++ src/main/java/baseball/domain/computer/Computer.java
@@ -1,38 +1,19 @@
 package baseball.domain.computer;
 
-import static baseball.domain.script.Script.GAME_START;
-
-import camp.nextstep.edu.missionutils.Randoms;
-import java.util.ArrayList;
 import java.util.List;
 
 public class Computer {
-    List<Integer> numbers;
+    private List<Integer> numbers;
 
     public Computer() {
-        List<Integer> numbers = new ArrayList<>();
-        while (numbers.size() < 3) {
-            int randomNumber = Randoms.pickNumberInRange(1, 9);
-            if (!numbers.contains(randomNumber)) {
-                numbers.add(randomNumber);
-            }
-        }
+
+    }
+
+    public void updateNumbers(List<Integer> numbers) {
         this.numbers = numbers;
     }
 
     public int getIndexNumbers(int i) {
-        return numbers.get(i);
-    }
-
-    public void resetNumbers() {
-        System.out.println(GAME_START.getMessage());
-        List<Integer> numbers = new ArrayList<>();
-        while (numbers.size() < 3) {
-            int randomNumber = Randoms.pickNumberInRange(1, 9);
-            if (!numbers.contains(randomNumber)) {
-                numbers.add(randomNumber);
-            }
-        }
-        this.numbers = numbers;
+        return this.numbers.get(i);
     }
 }
